# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ActorDetail": {
        "description": "Response for ``GET /api/v1/actors/{id}`` (plan D1 + D11).\n\nAll core ``ActorItem`` fields plus no linked-reports collection:\nper plan D11 out-of-scope ruling, ``ActorDetail`` does NOT\ntraverse ``report_codenames`` to surface reports that mention\nthis actor. That surface needs a dedicated endpoint whose filter\ncontract, pagination, and RBAC scope have not been locked yet —\ncarried to a later Phase 3 slice. The FE consequence: there is\nno \"recent reports for this actor\" panel on the actor detail\npage this PR.",
        "examples": [
          {
            "aka": [
              "APT38",
              "Hidden Cobra"
            ],
            "codenames": [
              "Andariel",
              "Bluenoroff"
            ],
            "description": "DPRK-attributed cyber espionage and financially motivated group",
            "id": 3,
            "mitre_intrusion_set_id": "G0032",
            "name": "Lazarus Group"
          }
        ],
        "properties": {
          "aka": {
            "items": {
              "type": "string"
            },
            "title": "Aka",
            "type": "array"
          },
          "codenames": {
            "items": {
              "type": "string"
            },
            "title": "Codenames",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mitre_intrusion_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Intrusion Set Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ActorDetail",
        "type": "object"
      },
      "ActorItem": {
        "description": "Single actor row in the ``/actors`` list view.\n\n``codenames`` flattens the ``codenames`` table join on\n``group_id``. Fields mirror the v1.0 §5.4 shape verbatim so the\nFE consumer lands without a rename pass.",
        "examples": [
          {
            "aka": [
              "APT38",
              "Hidden Cobra"
            ],
            "codenames": [
              "Andariel",
              "Bluenoroff"
            ],
            "description": "DPRK-attributed cyber espionage and financially motivated group",
            "id": 3,
            "mitre_intrusion_set_id": "G0032",
            "name": "Lazarus Group"
          }
        ],
        "properties": {
          "aka": {
            "items": {
              "type": "string"
            },
            "title": "Aka",
            "type": "array"
          },
          "codenames": {
            "items": {
              "type": "string"
            },
            "title": "Codenames",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mitre_intrusion_set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mitre Intrusion Set Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ActorItem",
        "type": "object"
      },
      "ActorListResponse": {
        "description": "Offset-paginated response for ``GET /api/v1/actors``.\n\nPlan D3 lock: actors stay on offset because the group count is\nsmall and sort-stable under inserts (``name ASC, id ASC`` per\nD11). ``total`` lets the FE compute remaining pages without a\nsecond ``HEAD``-style call.",
        "examples": [
          {
            "items": [
              {
                "aka": [
                  "APT38",
                  "Hidden Cobra"
                ],
                "codenames": [
                  "Andariel",
                  "Bluenoroff"
                ],
                "description": "DPRK-attributed cyber espionage and financially motivated group",
                "id": 3,
                "mitre_intrusion_set_id": "G0032",
                "name": "Lazarus Group"
              }
            ],
            "limit": 50,
            "offset": 0,
            "total": 12
          },
          {
            "items": [],
            "limit": 50,
            "offset": 0,
            "total": 0
          }
        ],
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActorItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "total"
        ],
        "title": "ActorListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Implementation-prep OpenAPI skeleton for the DPRK CTI platform.",
    "title": "DPRK CTI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/actors": {
      "get": {
        "description": "Returns the list of threat-actor groups (name, MITRE intrusion set id, aka aliases, description) together with their codenames. Offset pagination — actors is a small, sort-stable set; keyset pagination is used for /reports and /incidents (plan D3). Default sort is `name ASC, id ASC` (plan D11).",
        "operationId": "list_actors_endpoint_api_v1_actors_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "empty": {
                    "summary": "No groups seeded yet",
                    "value": {
                      "items": [],
                      "limit": 50,
                      "offset": 0,
                      "total": 0
                    }
                  },
                  "happy": {
                    "summary": "Three groups, first page",
                    "value": {
                      "items": [
                        {
                          "aka": [
                            "APT38",
                            "Hidden Cobra"
                          ],
                          "codenames": [
                            "Andariel",
                            "Bluenoroff"
                          ],
                          "description": "DPRK-attributed",
                          "id": 3,
                          "mitre_intrusion_set_id": "G0032",
                          "name": "Lazarus Group"
                        }
                      ],
                      "limit": 50,
                      "offset": 0,
                      "total": 3
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ActorListResponse"
                }
              }
            },
            "description": "One page of actors with offset metadata."
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Authenticated role is not analyst / researcher / policy / soc / admin"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "msg": "Input should be less than or equal to 200",
                      "type": "less_than_equal"
                    }
                  ]
                }
              }
            },
            "description": "Invalid query parameter (limit out of 1..200, offset negative). Plan D12 — no silent ignore of filter values."
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": "rate_limit_exceeded",
                  "message": "60 per 1 minute"
                }
              }
            },
            "description": "Rate limit exceeded — 60/min/user read bucket (plan D2). `Retry-After`, `X-RateLimit-Limit`, `X-RateLimit-Remaining` headers present. Body is the uniform JSON shape pinned by Group G (`api.rate_limit.rate_limit_exceeded_handler`)."
          }
        },
        "summary": "List threat actors (offset-paginated)",
        "tags": [
          "actors"
        ]
      }
    },
    "/api/v1/actors/{actor_id}": {
      "get": {
        "description": "Shallow-joined detail view of a single threat-actor group. Returns the core group fields plus the flat codenames list. Plan D11 excludes traversal into `report_codenames` — this endpoint does NOT surface reports that mention the actor. Returns 404 when the actor id is unknown.",
        "operationId": "get_actor_detail_endpoint_api_v1_actors__actor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "title": "Actor Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorDetail"
                }
              }
            },
            "description": "Actor detail with codenames (plan D1 + D11)."
          },
          "401": {
            "description": "Missing or invalid session cookie"
          },
          "403": {
            "description": "Role not analyst / researcher / policy / soc / admin"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "actor not found"
                }
              }
            },
            "description": "Actor id not found"
          },
          "422": {
            "description": "Invalid path parameter — non-integer actor_id."
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": "rate_limit_exceeded",
                  "message": "60 per 1 minute"
                }
              }
            },
            "description": "Rate limit exceeded — 60/min/user per-route bucket."
          }
        },
        "summary": "Get one threat actor (group) with its codenames",
        "tags": [
          "actors"
        ]
      }
    },
...
  }
}